Rasa-bot4/actions/requirements-actions.txt
pymongo
pymongo[srv]
importlib_metadata
python-engineio==4.3.0
python-socketio==5.5.0
python-dotenv
requests
datetime
